feat: currentContextPage일 때 JournalForm 작업
- 현재 페이지가 currentContext일때만 JournalForm이 보이도록 작업
- journal.sttyle.ts파일로  styledComponent분리작업

Relative to:#156
